Clean rotation-period extras and finalize Fourier core split
