--- docs/rotation_period_validation.md
@@ -1,121 +1,0 @@
-# Rotation Period Analysis and Validation
-
-## What Was Added
-
-`adam_core.photometry` now includes a rotation-period solver based on a high-order
-Fourier frequency search, along with detection wrappers and optional real-data
-regression fixtures.
-
-The main public entry points are:
-
-- `estimate_rotation_period(...)`
-- `build_rotation_period_observations_from_detections(...)`
-- `estimate_rotation_period_from_detections(...)`
-- `estimate_rotation_period_from_detections_grouped(...)`
-
-The observation model now supports an optional `session_id` field. When present,
-the solver fits within-filter session offsets in addition to the existing filter
-offsets, phase-function terms, and Fourier harmonics. This is important for
-real lightcurve archives where different nights or submitted lightcurve blocks
-often have different zero points.
-
-## Validation Strategy
-
-The project now uses two layers of validation:
-
-1. Synthetic unit tests
-
-- period recovery for a known signal
-- multi-filter offset handling
-- session-offset handling
-- single-peaked period doubling
-- outlier rejection
-- minimum-sufficient Fourier-order selection
-- detection-wrapper and grouped-wrapper behavior
-
-2. Optional real-data regression tests
-
-- legacy MPC-backed path retained as a heterogeneous stress test
-- new PDS-backed path added as the primary scientific regression path
-- LCDB supplies the expected published period
-- ALCDEF supplies the raw time-series lightcurve sessions
-
-The PDS regression suite is opt-in:
-
-```bash
-pdm run pytest --benchmark-skip src/adam_core/photometry/tests/test_rotation_period_pds.py -q --run-rotation-period-pds
-```
-
-## Current PDS Gold Fixtures
-
-The current gold set is:
-
-- `289 Nenetta`
-- `1011 Laodamia`
-- `886 Washingtonia`
-- `1282 Utopia`
-- `1323 Tugela`
-- `1627 Ivar`
-
-These fixtures include:
-
-- times
-- magnitudes
-- magnitude errors when present
-- filters
-- `session_id`
-- geometry arrays (`r_au`, `delta_au`, `phase_angle_deg`)
-- LCDB expected period and quality metadata
-- per-session metadata copied from ALCDEF for auditability
-
-## Why This Is Not Overfit
-
-The current set is not just a narrow fit to the sample objects.
-
-First, the solver mechanics are covered by synthetic tests that are independent
-of the archive fixtures. Those tests verify the behaviors that matter for the
-algorithm itself: harmonics, nuisance offsets, clipping, order selection, and
-the single-peak/double-peak ambiguity. The PDS cases are not the only tests
-exercising the implementation.
-
-Second, the expected periods come from LCDB while the photometry comes from
-ALCDEF. Those are separate archives with separate purposes, so the solver is
-recovering published periods from independent raw lightcurves rather than
-matching a target generated by the code under test.
-
-Third, the gold/challenge split is based on curation and recovery stability,
-not object identity. Objects only enter the gold set when the current solver
-recovers them tightly on a clean subset. Rougher cases stay in the challenge
-suite until the model and preprocessing prove they are stable enough to
-promote.
-
-The session-aware extension is still justified by the archive structure itself.
-ALCDEF metadata explicitly preserves lightcurve-block boundaries, and raw
-multi-session asteroid photometry commonly needs per-session zero-point
-freedom. That is a model change driven by the data model, not a special-case
-tweak for a single asteroid.
-
-## What We Know So Far
-
-With the current implementation and frozen PDS fixtures, the solver recovers:
-
-- `1011 Laodamia`: `5.173269 h` vs `5.172470 h` (`0.015%`)
-- `1282 Utopia`: `13.632138 h` vs `13.623000 h` (`0.067%`)
-- `1323 Tugela`: `19.526439 h` vs `19.500000 h` (`0.136%`)
-- `1627 Ivar`: `4.795401 h` vs `4.795000 h` (`0.008%`)
-- `289 Nenetta`: `6.916690 h` vs `6.902000 h` (`0.213%`)
-- `886 Washingtonia`: `8.999436 h` vs `9.001000 h` (`0.017%`)
-- `511 Davida` challenge case: `5.126542 h` vs `5.129700 h` (`0.062%`)
-- `702 Alauda` challenge case: `16.706116 h` vs `16.707200 h` (`0.006%`)
-
-That is materially tighter than the earlier MPC-backed validation path, which
-was dominated by heterogeneous photometric systematics.
-
-## Current Limits
-
-- The gold PDS suite now covers six objects, but it is still a curated subset of the archive.
-- The optional PDS suite is slower than the synthetic suite; the expanded
-  eight-fixture run took `5:35` in the latest local pass.
-- `511 Davida` and `702 Alauda` remain challenge cases with looser tolerances.
-- `433 Eros` remains excluded from the frozen suite because the tested subset
-  converged to an alias / wrong period under the current solver.
