--- src/adam_core/photometry/tests/test_rotation_period_real_data.py
@@ -1,78 +1,0 @@
-from __future__ import annotations
-
-from pathlib import Path
-
-import numpy as np
-import pyarrow as pa
-import pytest
-
-from adam_core.photometry.rotation_period_fourier import estimate_rotation_period
-from adam_core.photometry.rotation_period_types import RotationPeriodObservations
-from adam_core.time import Timestamp
-
-DATA_DIR = Path(__file__).parent / "data"
-
-REAL_FIXTURES: list[str] = sorted(
-    p.name for p in DATA_DIR.glob("rotation_period_fixture_*.npz")
-)
-if not REAL_FIXTURES:
-    REAL_FIXTURES = ["__NO_FIXTURES__"]
-
-
-def _scalar(value) -> object:
-    if hasattr(value, "as_py"):
-        return value.as_py()
-    arr = np.asarray(value)
-    if arr.shape == ():
-        return arr.item()
-    if arr.size != 1:
-        raise ValueError(f"Expected scalar-like value, got shape {arr.shape}")
-    return arr.reshape(-1)[0].item()
-
-
-@pytest.mark.parametrize("fixture_name", REAL_FIXTURES)
-def test_rotation_period_from_real_fixture(
-    fixture_name: str,
-    pytestconfig: pytest.Config,
-) -> None:
-    if not pytestconfig.getoption("--run-rotation-period-real-data"):
-        pytest.skip("Real-data rotation-period regression tests are opt-in.")
-    if fixture_name == "__NO_FIXTURES__":
-        pytest.skip("No real-data rotation-period fixtures found on disk.")
-
-    fx = np.load(DATA_DIR / fixture_name, allow_pickle=True)
-    observations = RotationPeriodObservations.from_kwargs(
-        time=Timestamp.from_iso8601(fx["time_iso"].astype(object).tolist(), scale="utc"),
-        mag=np.asarray(fx["mag_obs"], dtype=np.float64),
-        mag_sigma=pa.array(
-            np.asarray(fx["rmsmag"], dtype=np.float64),
-            mask=~np.isfinite(np.asarray(fx["rmsmag"], dtype=np.float64)),
-            type=pa.float64(),
-        ),
-        filter=fx["filter"].astype(object).tolist(),
-        session_id=(
-            fx["session_id"].astype(object).tolist()
-            if "session_id" in fx.files
-            else [None] * len(fx["time_iso"])
-        ),
-        r_au=np.asarray(fx["r_au"], dtype=np.float64),
-        delta_au=np.asarray(fx["delta_au"], dtype=np.float64),
-        phase_angle_deg=np.asarray(fx["phase_angle_deg"], dtype=np.float64),
-    )
-
-    result = estimate_rotation_period(
-        observations,
-        frequency_grid_scale=float(fx["frequency_grid_scale"][0]),
-        max_frequency_cycles_per_day=float(fx["max_frequency_cycles_per_day"][0]),
-        min_rotations_in_span=float(fx["min_rotations_in_span"][0]),
-    )
-    period_hours = float(_scalar(result.period_hours[0]))
-    expected_hours = float(fx["expected_period_hours"][0])
-    tolerance_fraction = float(fx["tolerance_fraction"][0])
-
-    rel_err = abs(period_hours - expected_hours) / expected_hours
-    assert rel_err <= tolerance_fraction, (
-        f"{fixture_name}: fitted {period_hours:.6f} h, expected {expected_hours:.6f} h, "
-        f"relative error {rel_err:.3%}, allowed {tolerance_fraction:.3%}. "
-        f"Source: {fx['source_title'][0]} ({fx['source_url'][0]})"
-    )
